refactor: 메서드 정렬 순서 변경
